backend/requirements.txt
streamlit
fastapi
uvicorn[standard]
requests
python-multipart
pypdf
python-docx
beautifulsoup4
lxml